fixed type cast

+++ src/matrix2d.cc
@@ -232,8 +232,8 @@ TMatrix2D::map(int inX, int inY, long *outX, long *outY) const
 {
   double x, y;
   x = inX; y=inY;
-  *outX = static_cast<int>(a11 * x + a12 * y + tx);
-  *outY = static_cast<int>(a21 * x + a22 * y + ty);
+  *outX = static_cast<long>(a11 * x + a12 * y + tx);
+  *outY = static_cast<long>(a21 * x + a22 * y + ty);
 }
 
 void
